--- src/specify_cli/commands/extension.py
@@ -1,2 +1,0 @@
-"""specify extension * commands — placeholder for future extraction."""
-from __future__ import annotations
