--- src/specify_cli/commands/integration.py
@@ -1,2 +1,0 @@
-"""specify integration * commands — placeholder for future extraction."""
-from __future__ import annotations
