--- src/specify_cli/commands/preset.py
@@ -1,2 +1,0 @@
-"""specify preset * commands — placeholder for future extraction."""
-from __future__ import annotations
