--- src/specify_cli/commands/workflow.py
@@ -1,2 +1,0 @@
-"""specify workflow * commands — placeholder for future extraction."""
-from __future__ import annotations
